% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{fit_model_func}
\alias{fit_model_func}
\title{Determine the Engine Used for Model Fitting}
\usage{
fit_model_func(model)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}
}
\value{
Function (\code{stats::lm} or \code{stats::glm}) corresponding to the input model's class.
}
\description{
This function checks the class of the input model and returns the appropriate
engine function (\code{stats::lm} or \code{stats::glm}), used for model fitting. This function is
utilized as a helper in various stability assessments.
}
\keyword{internal}
