% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstab.R
\name{gstab}
\alias{gstab}
\title{Stability Analysis for Models}
\usage{
gstab(model, ...)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Depends on the method called. If 'lm', a list with the original model
and the results of each stability check.
}
\description{
Generic function for performing a stability analysis on a model. It calls the
appropriate method based on the class of its argument.
}
\keyword{internal}
