% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{handle_boot_model_summaries}
\alias{handle_boot_model_summaries}
\title{Handle Bootstrap Model Summaries}
\usage{
handle_boot_model_summaries(boot_models, conf.int, conf.level)
}
\arguments{
\item{boot_models}{A list of bootstrapped model objects.}

\item{conf.int}{Logical, if \code{TRUE} confidence intervals for the estimates are
returned.}

\item{conf.level}{The confidence level to use for the confidence intervals.}
}
\value{
A list with two elements:
\itemize{
\item 'boot_mean': A data frame with the mean values for each model term, including
the estimate, standard error, proportion of significant p-values, and if
\code{conf.int} is \code{TRUE}, the confidence interval for the estimate.
\item 'boot_sd': A data frame with the standard deviations of the estimate and
standard error for each model term.
}
}
\description{
This function computes and returns summary statistics for a list of
bootstrapped model objects. If the input list is \code{NULL}, the function returns
a list with two \code{NULL} elements.
}
\keyword{internal}
