% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n20_seed376_bin}
\alias{n20_seed376_bin}
\title{Simulated Data: n20_seed376_bin}
\format{
\subsection{\code{n20_seed376_bin}}{

A data frame with 20 rows and 5 columns:
\describe{
\item{y}{Binomial dependent variable}
\item{x1, x2, x3}{Continuous independent variables}
\item{x4}{Factor independent variable}
}
}
}
\source{
Simulated data
}
\usage{
n20_seed376_bin
}
\description{
A data set simulated for stats::glm engine's unit testing and examples. This
data set is designed to be used with the family parameter of stats::glm set to
binomial(link = "logit").
}
\keyword{datasets}
