% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n20_seed500_lm}
\alias{n20_seed500_lm}
\title{Simulated Data: n20_seed500_lm}
\format{
\subsection{\code{n20_seed500_lm}}{

A data frame with 20 rows and 5 columns:
\describe{
\item{y}{Normally distributed dependent variable}
\item{x1, x2, x3}{Continuous independent variables}
\item{x4}{Factor independent variable}
}
}
}
\source{
Simulated data
}
\usage{
n20_seed500_lm
}
\description{
A data set simulated for stats::lm engine's unit testing and examples.
}
\keyword{datasets}
