% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n20_seed587_pois}
\alias{n20_seed587_pois}
\title{Simulated Data: n20_seed587_pois}
\format{
\subsection{\code{n20_seed587_pois}}{

A data frame with 20 rows and 3 columns:
\describe{
\item{y}{Poisson dependent variable}
\item{x1, x2}{Continuous independent variables}
}
}
}
\source{
Simulated data
}
\usage{
n20_seed587_pois
}
\description{
A data set simulated for stats::glm engine's unit testing and examples. This
data set is designed to be used with the family parameter of stats::glm set to
poisson(link = "log").
}
\keyword{datasets}
