% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{numerical_stability}
\alias{numerical_stability}
\title{Numerical Stability Assessment}
\usage{
numerical_stability(model, data, formula, ...)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}

\item{data}{A data frame containing the data used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{...}{Other parameters to be passed to the fitting function determined by
\code{fit_model_func}.}
}
\value{
The function returns the fitted model object obtained from the perturbed
data.
}
\description{
This function assesses the numerical stability of a fitted model by introducing
a small perturbation to the numerical variables in the dataset and comparing
the resulting model to the original. The results of the function can be used
to assess the robustness of a model to rounding errors and limited precision
computations.
}
\keyword{internal}
