% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstab_plot.R
\name{plot.gstab_sum}
\alias{plot.gstab_sum}
\title{Plotting Function for \code{gstab_sum} Objects}
\usage{
\method{plot}{gstab_sum}(x, conf.int = TRUE, ...)
}
\arguments{
\item{x}{A \code{gstab_sum} object, which is the output of the \code{summary.gstab}
function.}

\item{conf.int}{Logical. If TRUE, confidence intervals are added to the plots.
Default is TRUE.}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\value{
A list of ggplot2 objects, each representing a different stability
assessment.
}
\description{
This function creates a list of plots for each of the stability assessments
performed in the \code{gstab} function. The plots are intended to provide a
visual overview of the stability of the model estimates under various
conditions.
}
\examples{
n <- 20
set.seed(376)
data <- data.frame(y = 3*stats::rnorm(n) +5,
                   x1 = 3*stats::rnorm(n) +5 + stats::rnorm(n, 2, 0.3),
                   x2 = 2*stats::rnorm(n) + 1.5*stats::rnorm(n) + stats::rnorm(n, 1, 0.05))
formula <- y ~ x1 + x2
model <- stats::lm(formula, data = data)
stab_res <- gstab(model = model)
stab_sum <- summary(stab_res, conf.int = TRUE, conf.level = 0.95)
plot(stab_sum, conf.int = TRUE)
}
