% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotters.R
\name{plot_analytic_and_algebraic_stability}
\alias{plot_analytic_and_algebraic_stability}
\title{Create Analytic and Algebraic Stability Assessment Plot}
\usage{
plot_analytic_and_algebraic_stability(obj)
}
\arguments{
\item{obj}{A list object returned from the \code{analytic_and_algebraic_stability}
function, containing the condition numbers for L1 and Linf norms.}
}
\value{
A ggplot2 object displaying the condition numbers based on the L1 and
Linf norms, indicating the sensitivity of the model to changes in the
independent variables.
}
\description{
This function creates a ggplot object for the analytic and algebraic stability
assessment. It visualizes the comparison of condition numbers (kappa) based
on the L1 and Linf norms.
}
\keyword{internal}
