% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotters.R
\name{plot_data_selection_stability}
\alias{plot_data_selection_stability}
\title{Create Stability under Data Selection Assessment Plot}
\usage{
plot_data_selection_stability(obj, conf.int)
}
\arguments{
\item{obj}{A list object returned from the \code{stability_under_data_selection}
function, containing original model summary, and summaries for models fitted
on bootstrap data, no outlier data, and three stratified bootstrap datasets.}

\item{conf.int}{Logical variable. If \code{TRUE}, add error bars to the points in
the plot, representing confidence intervals.}
}
\value{
A list of ggplot2 objects of the comparison between original and
various data selected model estimates.
}
\description{
This function creates a list of ggplot objects for the stability under data
selection assessment. It generates plots comparing original model estimates
with estimates from models fitted on resampled data, data with outliers
removed, and stratified bootstrap data.
}
\keyword{internal}
