% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotters.R
\name{plot_numerical_stability}
\alias{plot_numerical_stability}
\title{Create Numerical Stability Assessment Plot}
\usage{
plot_numerical_stability(obj, conf.int)
}
\arguments{
\item{obj}{A list object returned from the \code{numerical_stability} function,
containing summaries for the original model and the model fitted on perturbed
data.}

\item{conf.int}{A number indicating the confidence interval level
(e.g., 0.95 for a 95\% confidence interval).}
}
\value{
A ggplot2 object of the comparison of estimates from the original
model and the perturbed model.
}
\description{
This function creates a ggplot object for the numerical stability assessment.
It visualizes the comparison of the original model estimates and the estimates
from the model fitted on data perturbed by a small amount of random noise.
}
\keyword{internal}
