% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotters.R
\name{plot_replication_stability}
\alias{plot_replication_stability}
\title{Create Replication Stability Assessment Plot}
\usage{
plot_replication_stability(obj, conf.int)
}
\arguments{
\item{obj}{A list object returned from the \code{replication_stability} function,
containing original model summary, and summaries for new model and bootstrap
models.}

\item{conf.int}{Logical variable. If \code{TRUE}, add error bars to the points in
the plot, representing confidence intervals.}
}
\value{
A list containing ggplot2 objects. The list includes \code{new} which is a
ggplot object of the comparison between original and new model estimates
(if new data was provided), and \code{boot} which is a ggplot object of the
comparison between original and bootstrap mean estimates (if bootstrap
resampling was performed).
}
\description{
This internal function creates ggplot objects for the replication stability
assessment. It generates two separate plots, one comparing original model
estimates with new model estimates, and another comparing original model
estimates with bootstrap mean estimates.
}
\keyword{internal}
