% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotters.R
\name{plot_statistical_stability}
\alias{plot_statistical_stability}
\title{Create Statistical Stability Assessment Plot}
\usage{
plot_statistical_stability(obj, conf.int)
}
\arguments{
\item{obj}{A list object returned from the \code{statistical_stability} function,
containing original model summary, and summaries for models fitted on noisy
data and permuted noisy data.}

\item{conf.int}{Logical variable. If \code{TRUE}, add error bars to the points in
the plot, representing confidence intervals.}
}
\value{
A ggplot2 object of the comparison between original, noisy, and
permuted noisy model estimates.
}
\description{
This internal function creates a ggplot object for the statistical stability
assessment. It generates a plot comparing original model estimates with
estimates from models fitted on data with added noise and permuted noise.
}
\keyword{internal}
