% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plotters.R
\name{plot_technique_stability}
\alias{plot_technique_stability}
\title{Create Stability Under Selection of Technique Assessment Plot}
\usage{
plot_technique_stability(obj, conf.int)
}
\arguments{
\item{obj}{A list object returned from the 'stability_under_selection_of_technique'
function, containing the model summaries for the original and robust regression
models.}

\item{conf.int}{Logical. If TRUE, confidence intervals are added to the plot.}
}
\value{
A ggplot2 object showing the model estimates from the original and
robust regression models, demonstrating the stability of the model under
different fitting techniques. If the robust model did not converge, a NULL
object is returned.
}
\description{
This function generates a ggplot object that visualizes the comparison between
the original model and a robust regression model, assessing the stability of
the model under different fitting techniques.
}
\keyword{internal}
