% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{replication_stability}
\alias{replication_stability}
\title{Perform Replication Stability Assessment}
\usage{
replication_stability(model, data, formula, new_data = NULL, nboot = NULL, ...)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}

\item{data}{A data frame containing the data used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{new_data}{An optional data frame to be used for replication stability
assessment. If provided, the model will be fitted on this new data set.}

\item{nboot}{An optional integer specifying the number of bootstrap resamples
to use for replication stability assessment.}

\item{...}{Additional arguments to be passed to the \code{fit_model} function.}
}
\value{
A list containing \code{new_model} fitted on new_data (if provided),
and \code{boot_models} which is a list of models fitted on bootstrap samples
(if nboot is provided).
}
\description{
This function implements the replication stability assessment. It fits the
original model to a new data set (if provided) and performs bootstrap resampling
(if specified) to assess the stability of the model across different samples.
}
\keyword{internal}
