% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_explainer.R
\name{stab_explainer.gstab_glm}
\alias{stab_explainer.gstab_glm}
\title{Stability Explainer for Generalized Linear Models}
\usage{
\method{stab_explainer}{gstab_glm}(object)
}
\arguments{
\item{object}{A \code{gstab_glm} object.}
}
\value{
A list of explanations and interpretations for each type of stability
analysis for generalized linear models.
}
\description{
Provides explanations and interpretations for each type of stability analysis
for generalized linear models.
}
\examples{
# To be provided...

}
\seealso{
\code{\link[base]{UseMethod}}
}
