% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_explainer.R
\name{stab_explainer.gstab_lm}
\alias{stab_explainer.gstab_lm}
\title{Stability Explainer for Linear Models}
\usage{
\method{stab_explainer}{gstab_lm}(object)
}
\arguments{
\item{object}{A \code{gstab_lm} object.}
}
\value{
A list of explanations and interpretations for each type of stability
analysis for linear models.
}
\description{
Provides explanations and interpretations for each type of stability analysis
for linear models.
}
\keyword{internal}
