% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_explainer.R
\name{stab_explainer}
\alias{stab_explainer}
\title{Stability Explainer}
\usage{
stab_explainer(object)
}
\arguments{
\item{object}{A \code{gstab} object.}
}
\value{
A list of explanations and interpretations for each type of stability
analysis.
}
\description{
Provides explanations and interpretations for each type of stability analysis.
}
\keyword{internal}
