% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{stability_app}
\alias{stability_app}
\title{Stability App}
\usage{
stability_app()
}
\value{
A Shiny web application.
}
\description{
This Shiny web application allows users to perform stability assessments on
data loaded from a CSV file. Users can specify the formula for the model,
choose the model type, and specify various other parameters.
The app will then perform the stability assessment and display the results as
tables, plots, and explanations.
}
\examples{
# Only run this example in interactive R sessions
if(interactive()){
  stability_app()
}
}
\keyword{interactive}
