% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_assessment.R
\name{stability_assessment}
\alias{stability_assessment}
\title{Stability Assessment}
\usage{
stability_assessment(
  data,
  formula,
  engine,
  new_data = NULL,
  nboot = NULL,
  variable_to_remove = NULL,
  variable_of_interest = NULL,
  conf.int = TRUE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{A data frame containing the data to be used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{engine}{A function for fitting the model. This could be a function like
\code{stats::lm} or \code{stats::glm}.}

\item{new_data}{An optional data frame to be used for replication stability
assessment. If provided, the model will be fitted on this new data set and the
results compared with the original model.}

\item{nboot}{An optional integer specifying the number of bootstrap resamples
to use for replication stability assessment.}

\item{variable_to_remove}{An optional character string specifying a variable
to be removed for stability under model selection assessment.}

\item{variable_of_interest}{An optional character string specifying a variable
of interest for backward selection in stability under model selection assessment.}

\item{conf.int}{Logical variable. If TRUE, include a confidence interval in
the summary. Defaults to TRUE.}

\item{conf.level}{Confidence level for confidence interval when conf.int is
TRUE. Defaults to 0.95.}

\item{...}{Additional arguments to be passed to the \code{engine} function.}
}
\value{
A list containing the results of the stability assessment. The list
includes the original \code{gstab} object, the summary of the \code{gstab} object, the
plots of the \code{gstab_sum} object, and the explanations of the \code{gstab} object.
}
\description{
Performs a stability assessment of a model based on the framework presented by
Michailides and de Leeuw (1998). The stability assessment considers various
forms of stability, including replication stability, statistical stability,
stability under data selection, stability under model selection, numerical
stability, analytic and algebraic stability, and stability under selection of
technique.
}
\examples{
formula <- y ~ x1 + x2
stability_assessment(data = n20_seed376_lm,
                     formula = formula,
                     engine = stats::lm)
}
\references{
Michailides, G., & de Leeuw, J. (1998). The Gifi system for nonlinear multivariate analysis. eScholarship, University of California, Los Angeles. \url{https://escholarship.org/uc/item/0789f7d3}
}
