% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{stability_under_data_selection}
\alias{stability_under_data_selection}
\title{Perform Stability Under Data Selection Assessment}
\usage{
stability_under_data_selection(model, data, formula, ...)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}

\item{data}{A data frame containing the data used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{...}{Additional arguments to be passed to the \code{fit_model} function.}
}
\value{
A list containing \code{bootstrap_model} fitted on resampled data,
\code{no_outlier_model} fitted on data with outliers removed, and \code{strata_boot_models}
list of models fitted on stratified bootstrap data.
}
\description{
This function implements the stability under data selection assessment. It
fits the specified model on three sets of data: resampled data, data with
outliers removed, and stratified bootstrap data, thereby assessing the model's
sensitivity to outliers and sampling variability.
}
\keyword{internal}
