% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{stability_under_selection_of_technique}
\alias{stability_under_selection_of_technique}
\title{Stability Under Selection of Technique Assessment}
\usage{
stability_under_selection_of_technique(model, data, formula, ...)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}

\item{data}{A data frame containing the data used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{...}{Other parameters passed to \code{robustbase::lmrob} or \code{robustbase::glmrob}.}
}
\value{
A fitted robust regression model object (\code{robustbase::lmrob} or
\code{robustbase::glmrob}), or NULL if the robust regression model could not be
fitted.
}
\description{
This function assesses the stability of a fitted model under the application
of different techniques. It fits a robust regression model using the same
specification as the original model and compares the results.
}
\keyword{internal}
