% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{statistical_stability}
\alias{statistical_stability}
\title{Perform Statistical Stability Assessment}
\usage{
statistical_stability(model, data, formula, nf = 0.05, ...)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}

\item{data}{A data frame containing the data used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{nf}{Noise factor as a percentage. For non-binomial families, this is used to
determine the standard deviation of the noise added to the transformed response
variable. For binomial families with a binary response, this is used to determine
the fraction of the response variable to flip (i.e., change 0 to 1 or 1 to 0).}

\item{...}{Additional arguments to be passed to the \code{fit_model} function.}
}
\value{
A list containing \code{noisy_model} fitted on the data with added random noise
and \code{permuted_noisy_model} fitted on the data with added permuted noise.
}
\description{
This function implements the statistical stability assessment by fitting the
original model to the data with added random noise and permuted noise. It takes into
account the family of the model (i.e., Gaussian, gamma, inverse Gaussian, Poisson,
quasi-Poisson, binomial, quasi-binomial), and assesses the model's sensitivity to
random variations in the data.
}
\details{
The function first determines the family of the model and applies the appropriate
transformation to the response variable. For Gaussian, gamma, and inverse Gaussian
families, it uses an identity or log transformation; for Poisson and quasi-Poisson,
a log transformation; and for binomial and quasi-binomial, a logit transformation
or a binary flip depending on whether the response variable is a proportion or binary.

If the response variable is binary (determined using a 'success/failure' flag), the
function flips a fraction of the response variable (changes from 0 to 1 or from 1 to 0).
For non-binary variables, it adds normally distributed noise to the transformed response.
After adding the noise, the inverse of the transformation is applied to the noisy response
to bring it back to the original scale.
}
\keyword{internal}
