% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstab_summary.R
\name{summary.gstab_lm}
\alias{summary.gstab_lm}
\title{Summary of Stability Analysis for Linear Models}
\usage{
\method{summary}{gstab_lm}(object, conf.int, conf.level, ...)
}
\arguments{
\item{object}{An object of class 'gstab_lm'.}

\item{conf.int}{Logical variable. If TRUE, include a confidence interval in
the summary.}

\item{conf.level}{Confidence level for confidence interval when conf.int is
TRUE.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with a summary of the original model and summaries of each stability check.
}
\description{
This function generates a summary of the stability analysis performed by 'gstab.lm'. It summarizes the original model and each stability check, returning a list of these summaries.
}
\keyword{internal}
