% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{summary_analytic_and_algebraic_stability}
\alias{summary_analytic_and_algebraic_stability}
\title{Summary of Analytic and Algebraic Stability}
\usage{
summary_analytic_and_algebraic_stability(obj)
}
\arguments{
\item{obj}{A list object returned by the \code{analytic_and_algebraic_stability}
function, containing the L1 norm and Linf norm condition numbers of the model.}
}
\value{
The function directly returns the input argument \code{obj}, which is a list
containing the L1 norm and Linf norm condition numbers of the model.
}
\description{
This function provides the summary for the analytic and algebraic stability
of a fitted model, calculated by the \code{analytic_and_algebraic_stability} function.
}
\keyword{internal}
