% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{summary_model_selection_stability}
\alias{summary_model_selection_stability}
\title{Summary Statistics for Stability Under Model Selection}
\usage{
summary_model_selection_stability(obj)
}
\arguments{
\item{obj}{A list object returned by \code{stability_under_model_selection}.}
}
\value{
A list of tidy tibbles for each component of the input list. Each tibble
contains summary statistics for the respective likelihood ratio test. This
includes degrees of freedom, test statistic value, and p-value, among other
things. If any error occurs during the tidying process, a warning is suppressed,
and the raw (untidied) result is returned.
}
\description{
This function extracts and tidies the summary statistics for each model
produced by the \code{stability_under_model_selection} function. The summary
statistics include likelihood ratio test results for each model comparison:
toggling the intercept, removing a specific variable, and removing the least
useful variable.
}
\keyword{internal}
