% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{summary_numerical_stability}
\alias{summary_numerical_stability}
\title{Summary Statistics for Numerical Stability}
\usage{
summary_numerical_stability(obj, conf.int, conf.level)
}
\arguments{
\item{obj}{A fitted model object returned by the \code{numerical_stability} function.}

\item{conf.int}{Logical indicating whether to compute confidence intervals for
the parameter estimates.}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int = TRUE}.}
}
\value{
A tibble summarizing the coefficients, standard errors, statistical
test results, and (optionally) confidence intervals for the coefficients of
the model. Each row corresponds to an individual term in the regression model's formula.
}
\description{
This function provides summary statistics for the model fitted on perturbed
data obtained by the \code{numerical_stability} function. The statistics summarize
the effects of small perturbations in the numerical variables on the model fit,
thereby assessing the model's numerical stability.
}
\keyword{internal}
