% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{summary_replication_stability}
\alias{summary_replication_stability}
\title{Summarize Results of Replication Stability Assessment}
\usage{
summary_replication_stability(obj, conf.int, conf.level)
}
\arguments{
\item{obj}{A list containing \code{new_model} fitted on new_data (if provided),
and \code{boot_models} which is a list of models fitted on bootstrap samples
(if nboot is provided).}

\item{conf.int}{Logical, if \code{TRUE} confidence intervals for the estimates are
returned.}

\item{conf.level}{The confidence level to use for the confidence intervals.}
}
\value{
A list with two elements:
\itemize{
\item 'new_model': A tidy data frame summarizing the model fitted on new_data
(if provided).
\item 'boot_models': A list of summarized bootstrap models, each as a tidy data
frame, (if bootstrap models were provided in 'obj').
}
}
\description{
This function summarizes the results of a replication stability assessment.
It creates a tidy data frame for the model fitted on new data (if provided),
and summarizes the bootstrap models (if available).
}
\keyword{internal}
