% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{summary_statistical_stability}
\alias{summary_statistical_stability}
\title{Summarize Results of Statistical Stability Assessment}
\usage{
summary_statistical_stability(obj, conf.int, conf.level)
}
\arguments{
\item{obj}{A list containing \code{noisy_model} fitted on the data with added
random noise and \code{permuted_noisy_model} fitted on the data with permuted noise.}

\item{conf.int}{Logical, if \code{TRUE} confidence intervals for the estimates are
returned.}

\item{conf.level}{The confidence level to use for the confidence intervals.}
}
\value{
A list with two elements:
\itemize{
\item 'noisy_model': A tidy data frame summarizing the model fitted on data with
added random noise.
\item 'permuted_noisy_model': A tidy data frame summarizing the model fitted on
data with permuted noise.
}
}
\description{
This function summarizes the results of a statistical stability assessment.
It creates tidy data frames for the model fitted on data with added random
noise and data with permuted noise.
}
\keyword{internal}
