% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{summary_technique_stability}
\alias{summary_technique_stability}
\title{Summary of Stability Under Selection of Technique}
\usage{
summary_technique_stability(obj, conf.int, conf.level)
}
\arguments{
\item{obj}{A fitted robust regression model object (\code{robustbase::lmrob} or
\code{robustbase::glmrob}), or NULL if the robust regression model could not be
fitted.}

\item{conf.int}{Logical indicating whether or not to compute confidence interval
around coefficients.}

\item{conf.level}{Numeric scalar indicating the confidence level to be used
for confidence interval calculation.}
}
\value{
A tidy data frame summarizing the robust regression model, or NULL
if an error occurs during tidying.
}
\description{
This function provides a summary of the stability under the selection of
technique of a fitted model. It generates the summary from a model object
obtained from the \code{stability_under_selection_of_technique} function.
}
\keyword{internal}
