% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_summarizers.R
\name{tidy_list_elements}
\alias{tidy_list_elements}
\title{Apply broom::tidy on List Elements}
\usage{
tidy_list_elements(list_elements, conf.int, conf.level)
}
\arguments{
\item{list_elements}{A list of model objects for which to extract statistics.}

\item{conf.int}{Logical, if \code{TRUE} confidence intervals for the estimates are
returned.}

\item{conf.level}{The confidence level to use for the confidence intervals.}
}
\value{
A list of tibbles with model statistics, same length as \code{list_elements}.
If \code{broom::tidy} operation fails on a model object, the corresponding list
element will be \code{NULL}.
}
\description{
This function applies the \code{broom::tidy} function to each element of a list,
extracting relevant statistics for each model object in the list. If the
\code{broom::tidy} operation fails on a model object, the function catches the
error, prints a warning, and returns \code{NULL} for that particular model object.
}
\keyword{internal}
