% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstab.R
\name{gstab.lm}
\alias{gstab.lm}
\title{Stability Analysis for Linear Models}
\usage{
\method{gstab}{lm}(
  model,
  new_data = NULL,
  nboot = NULL,
  variable_to_remove = NULL,
  variable_of_interest = NULL,
  ...
)
}
\arguments{
\item{model}{An object of class "lm": a fitted linear model.}

\item{new_data}{An optional data frame in which to look for variables with which
to predict. If omitted, the fitted values are used.}

\item{nboot}{Number of bootstrap resamples to perform in the replication
stability test. If NULL (default), no resampling is done.}

\item{variable_to_remove}{Variable to remove for the stability under model
selection test. If NULL (default), no variable is removed.}

\item{variable_of_interest}{Variable to consider for backward selection in the
stability under model selection test. If NULL (default), no variable is considered.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with the original model and the results of each stability check.
}
\description{
This function performs a stability analysis on a linear model (of class 'lm'),
checking for different types of stabilities as defined by Michailides and de
Leeuw (1998). The stability tests include replication stability, statistical
stability, stability under data selection, stability under model selection,
numerical stability, analytic and algebraic stability, and stability under
selection of technique. The output is a list with the original model and the
results of each stability check.
}
\examples{
# Generating example data
n <- 20
set.seed(376)
data <- data.frame(y = 3*stats::rnorm(n) +5,
                   x1 = 3*stats::rnorm(n) +5 + stats::rnorm(n, 2, 0.3),
                   x2 = 2*stats::rnorm(n) + 1.5*stats::rnorm(n) + stats::rnorm(n, 1, 0.05))
formula <- y ~ x1 + x2

# Fit the model
model <- stats::lm(formula, data = data)

# Performing stability analysis
gstab(model = model, nboot = 100)
}
\references{
Michailides, G., & de Leeuw, J. (1998). The Gifi system for nonlinear multivariate
analysis. eScholarship, University of California, Los Angeles.
https://escholarship.org/uc/item/0789f7d3
}
