% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{statistical_stability}
\alias{statistical_stability}
\title{Perform Statistical Stability Assessment}
\usage{
statistical_stability(model, data, formula, ...)
}
\arguments{
\item{model}{A fitted model object, either of class \code{lm} or \code{glm}.}

\item{data}{A data frame containing the data used for model fitting.}

\item{formula}{A formula describing the model to be fitted.}

\item{...}{Additional arguments to be passed to the \code{fit_model} function.}
}
\value{
A list containing \code{noisy_model} fitted on the data with added random noise
and \code{permuted_noisy_model} fitted on the data with added permuted noise.
}
\description{
This function implements the statistical stability assessment. It fits the
original model to the data with added random noise and with permuted noise, thereby
assessing the model's sensitivity to random variations in the data.
}
\keyword{internal}
