% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstab_summary.R
\name{summary.gstab_lm}
\alias{summary.gstab_lm}
\title{Summary of Stability Analysis for Linear Models}
\usage{
\method{summary}{gstab_lm}(object, conf.int, conf.level, ...)
}
\arguments{
\item{object}{An object of class 'gstab_lm'.}

\item{conf.int}{Logical variable. If TRUE, include a confidence interval in
the summary.}

\item{conf.level}{Confidence level for confidence interval when conf.int is
TRUE.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with a summary of the original model and summaries of each stability check.
}
\description{
This function generates a summary of the stability analysis performed by 'gstab.lm'. It summarizes the original model and each stability check, returning a list of these summaries.
}
\examples{
# Assuming that `model` is an object of class 'gstab_lm'
n <- 20
set.seed(376)
data <- data.frame(y = 3*stats::rnorm(n) +5,
                   x1 = 3*stats::rnorm(n) +5 + stats::rnorm(n, 2, 0.3),
                   x2 = 2*stats::rnorm(n) + 1.5*stats::rnorm(n) + stats::rnorm(n, 1, 0.05))

formula <- y ~ x1 + x2
model <- stats::lm(formula, data = data)

stab_res <- gstab(model = model, nboot = 100)
summary(stab_res, conf.int = TRUE, conf.level = 0.95)
}
\seealso{
\code{\link{gstab}} for performing the stability analysis,
}
